requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
discord.py==2.3.2
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
psycopg2-binary==2.9.9
pytz==2023.3.post1
shortuuid==1.0.11
sqlparse==0.4.4
typing_extensions==4.8.0
yarl==1.9.3
